refactor: destroy follows in a job
