+++ decidim-admin/app/commands/decidim/admin/destroy_participatory_space_private_user.rb
@@ -21,32 +21,12 @@ module Decidim
 
         return unless space.private_space?
 
-        return if space.respond_to?("is_transparent") && space.is_transparent?
+        return if space.respond_to?(:is_transparent) && space.is_transparent?
 
-        ids = []
-        follows = Decidim::Follow.where(user: resource.decidim_user_id)
-        ids << find_space_follow_id(follows, resource, space)
-        ids << find_children_follows_ids(follows, space)
-        follows.where(id: ids.flatten.compact).destroy_all if ids.present?
-      end
+        decidim_user_id = resource.decidim_user_id
+        privatable_to_type = resource.privatable_to_type
 
-      def find_space_follow_id(follows, resource, space)
-        follows.where(decidim_followable_type: resource.privatable_to_type)
-               .where(decidim_followable_id: space.id)
-               &.first&.id
-      end
-
-      def find_children_follows_ids(follows, space)
-        follows.map do |follow|
-          object = find_object_followed(follow).presence
-          next unless object.respond_to?("decidim_component_id")
-
-          follow.id if space.components.ids.include?(object.decidim_component_id)
-        end
-      end
-
-      def find_object_followed(follow)
-        follow.decidim_followable_type.constantize.find(follow.decidim_followable_id)
+        DestroyPrivateUsersFollowsJob.perform_later(decidim_user_id, privatable_to_type, space)
       end
     end
   end
